refactor: User DTO 생성 및 변경 및 적용 (#449)
* FEAT: move user.model.ts and add PrivateUser interface

- User/user.model.ts -> DTO/user.model.ts

* REFACT: move users.type.ts 's Meta interface file

* REFACTOR: Private user extends User
